refactor and rename for clarity
